requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
python-multipart
openai
sqlalchemy
psycopg2-binary
alembic
python-dotenv